evol ACL #180

+++ src/AppBundle/Controller/Rest/WatcherController.php
@@ -2,6 +2,7 @@
 
 namespace AppBundle\Controller\Rest;
 
+use AppBundle\Service\ApplicationService;
 use AppBundle\Service\WatcherService;
 use ATS\CoreBundle\Controller\Rest\RestControllerTrait;
 use Symfony\Component\Routing\Annotation\Route;
@@ -11,24 +12,47 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\HttpKernel\Exception\HttpException;
 use AppBundle\Exception\DuplicateApplicationNameException;
+use AppBundle\Document\AccessLevel;
 
 class WatcherController extends Controller
 {
     use RestControllerTrait;
+
+    public function hasUserPermission(ApplicationService $applicationService, $app, $envId){
+        $user = $this->getUser();
+        if($applicationService->userHasPermission(
+            $app, 
+            $user, 
+            $envId, 
+            array(AccessLevel::ADMIN, AccessLevel::EDITOR))){
+            return true;
+        } else {
+            return false;
+        }
+    }
 
     /**
      * @Route("/add", methods="POST")
      *
      * @param Request        $request
      * @param WatcherService $watcherService
+     * @param ApplicationService $applicationService
      *
      * @return Response
      */
-    public function saveOrUpdateAction(Request $request, WatcherService $watcherService) {
+    public function saveOrUpdateAction(
+        Request $request, 
+        WatcherService $watcherService,
+        ApplicationService $applicationService) {
         try {
             $data = $request->getContent();
-            $watcher = $watcherService->saveOrUpdate($data);
-            return $this->renderResponse($watcher, Response::HTTP_OK, []);
+            $_data = json_decode($data, true);
+            if($this->hasUserPermission($applicationService, $_data["appId"], $_data["envId"])){
+                $watcher = $watcherService->saveOrUpdate($data);
+                return $this->renderResponse($watcher, Response::HTTP_OK, []);
+            } else {
+                return $this->exception(['message' => "You dont have rights permissions"], 400);
+            }
         }
         catch (DuplicateApplicationNameException $e) {
             return $this->renderResponse(['message' => $e->getMessage()], Response::HTTP_NOT_ACCEPTABLE);
@@ -43,14 +67,23 @@ class WatcherController extends Controller
      *
      * @param Request        $request
      * @param WatcherService $watcherService
+     * @param ApplicationService $applicationService
      *
      * @return Response
      */
-    public function listAction(Request $request, WatcherService $watcherService) {
+    public function listAction(
+        Request $request,
+        WatcherService $watcherService,
+        ApplicationService $applicationService) {
         try {
             $data = $request->getContent();
-            $watcher = $watcherService->list(json_decode($data, true));
-            return $this->renderResponse($watcher, Response::HTTP_OK, []);
+            $_data = \json_decode($data, true);
+            if($this->hasUserPermission($applicationService, $_data["appId"], $_data["envId"])){
+                $watcher = $watcherService->list(json_decode($data, true));
+                return $this->renderResponse($watcher, Response::HTTP_OK, []);
+            } else {
+                return $this->exception(['message' => "You dont have rights permissions"], 400);
+            }
         }
          catch (\Exception $e) {
             return $this->exception($e->getMessage(), 400);
@@ -58,18 +91,29 @@ class WatcherController extends Controller
     }
 
     /**
-     * @Route("/{id}/delete", methods="DELETE")
+     * @Route("/{id}/delete", methods="POST")
      *
      * @param Request        $request
      * @param WatcherService $watcherService
+     * @param ApplicationService $applicationService
      * @param string $id
      *
      * @return Response
      */
-    public function deleteAction(Request $request, WatcherService $watcherService, $id) {
+    public function deleteAction(
+        Request $request, 
+        WatcherService $watcherService,
+        ApplicationService $applicationService,
+        $id) {
         try {
-            $watcher = $watcherService->delete($id);
-            return $this->renderResponse($watcher, Response::HTTP_OK, []);
+            $data = $request->getContent();
+            $_data = \json_decode($data, true);
+            if($this->hasUserPermission($applicationService, $_data["appId"], $_data["envId"])){
+                $watcher = $watcherService->delete($id);
+                return $this->renderResponse($watcher, Response::HTTP_OK, []);
+            } else {
+                return $this->exception(['message' => "You dont have rights permissions"], 400);
+            }
         }
          catch (\Exception $e) {
             return $this->exception($e->getMessage(), 400);
@@ -81,14 +125,26 @@ class WatcherController extends Controller
      *
      * @param Request        $request
      * @param WatcherService $watcherService
+     * @param ApplicationService $applicationService
      * @param string $id
      *
      * @return Response
      */
-    public function executeAction(Request $request, WatcherService $watcherService, $id) {
+    public function executeAction(
+        Request $request, 
+        WatcherService $watcherService,
+        ApplicationService $applicationService,
+         $id) {
         try {
-            $watcher = $watcherService->execute($id);
-            return $this->renderResponse($watcher, Response::HTTP_OK, []);
+            $data = $request->getContent();
+            $_data = \json_decode($data, true);
+            if($this->hasUserPermission($applicationService, $_data["appId"], $_data["envId"])){
+                $watcher = $watcherService->execute($id);
+                return $this->renderResponse($watcher, Response::HTTP_OK, []);
+            } else {
+                return $this->exception(['message' => "You dont have rights permissions"], 400);
+            }
+
         }
          catch (\Exception $e) {
             return $this->exception($e->getMessage(), 400);
